service-workers/service-worker/waiting.https.html

<!DOCTYPE html>
<title>ServiceWorker: navigator.serviceWorker.waiting</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>

const SCRIPT = 'resources/empty-worker.js';
const SCOPE = 'resources/blank.html';

promise_test(async t => {
  await service_worker_unregister(t, SCOPE);
  const frame = await with_iframe(SCOPE);
  const registration =
      await navigator.serviceWorker.register(SCRIPT, {scope: SCOPE});
  await wait_for_state(t, registration.installing, 'installed');
  const controller = frame.contentWindow.navigator.serviceWorker.controller;

  assert_equals(controller, null, 'controller');
  assert_equals(registration.active, null, 'registration.active');
  assert_equals(registration.waiting.state, 'installed',
                'registration.waiting');
  assert_equals(registration.installing, null, 'registration.installing');
  frame.remove();
  await service_worker_unregister_and_done(t, SCOPE);
}, 'waiting is set after installation');

// Tests that the ServiceWorker objects returned from waiting attribute getter
// that represent the same service worker are the same objects.
promise_test(async t => {
  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  const r = await navigator.serviceWorker.getRegistration(SCOPE);
  assert_equals(registration.waiting, r.waiting, 'getRegistration ' +
                'should return the same object');
  await service_worker_unregister_and_done(t, SCOPE);
}, 'The ServiceWorker objects returned from waiting attribute getter that ' +
   'represent the same service worker are the same objects');
</script>
</body>
